chore: reorganizing files a bit
