+++ pyannote/audio/interactive/common/commands.png
