+++ pyannote/audio/interactive/common/controller.js
