+++ pyannote/audio/interactive/common/utils.py
