+++ pyannote/audio/interactive/common/wavesurfer.js
