+++ pyannote/audio/interactive/diff/controller.js
