+++ setup.cfg
@@ -62,9 +62,9 @@ console_scripts =
     pyannote-audio-eval=pyannote.audio.cli.evaluate:evaluate
 
 prodigy_recipes =
-    pyannote.audio = pyannote.audio.interactive.recipes.pipeline:pipeline
-    pyannote.review = pyannote.audio.interactive.recipes.review:review
-    pyannote.diff = pyannote.audio.interactive.recipes.diff:diff
+    pyannote.audio = pyannote.audio.interactive.pipeline.recipe:pipeline
+    pyannote.review = pyannote.audio.interactive.review.recipe:review
+    pyannote.diff = pyannote.audio.interactive.diff.recipe:diff
 
 [test]
 # py.test options when running `python setup.py test`
